web-framework/requirements.txt
asyncio-mqtt==0.16.2
firebase-admin==6.4.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-MQTT==1.1.1
httplib2==0.22.0
keras==3.0.5
mariadb==1.1.10
numpy==1.26.4
paho-mqtt==1.6.1
schedule==1.2.1
tensorflow==2.16.1
gunicorn==21.2.0
schedule==1.2.1
python-dotenv==1.0.1
requests==2.31.0
python-weather==2.0.1
